# fullnode/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.Res:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  mparam.AddOauth2:
    properties:
      auth_url:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      name:
        enum:
        - github
        - facebook
        - google
        type: string
      redirect_url:
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
    required:
    - auth_url
    - client_id
    - client_secret
    - name
    - redirect_url
    - scopes
    - token_url
    type: object
  mparam.AddRelay:
    properties:
      host:
        description: api.github.com
        type: string
      name:
        type: string
      out_port:
        description: "443"
        type: integer
      port:
        description: "443"
        type: integer
    required:
    - host
    - name
    - out_port
    - port
    type: object
...
  mparam.EditOauth2:
    properties:
      auth_url:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      id:
        type: integer
      name:
        enum:
        - github
        - facebook
        - google
        type: string
      redirect_url:
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
    required:
    - auth_url
    - client_id
    - client_secret
    - id
    - name
    - redirect_url
    - scopes
    - token_url
    type: object
  mparam.EditRelay:
    properties:
      host:
        description: api.github.com
        type: string
      id:
        type: integer
      name:
        type: string
      out_port:
        description: "443"
        type: integer
      port:
        description: "443"
        type: integer
    required:
    - host
    - id
    - name
    - out_port
    - port
    type: object
...
host: 127.0.0.1:80
info:
  contact: {}
  description: This is ZAManager api list.
  title: ZAManager API
  version: 1.0.0
paths:
...
  /access/relay:
    get:
      description: 获取ZTA的relay
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: RelayList
      tags:
      - ZTA
    post:
      consumes:
      - application/json
      description: 新增ZTA的relay
      parameters:
      - description: 新增ZTA的relay
        in: body
        name: Relay
        required: true
        schema:
          $ref: '#/definitions/mparam.AddRelay'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: AddRelay
      tags:
      - ZTA
    put:
      consumes:
      - application/json
      description: 修改ZTA的relay
      parameters:
      - description: 修改ZTA的relay
        in: body
        name: Relay
        required: true
        schema:
          $ref: '#/definitions/mparam.EditRelay'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: EditRelay
      tags:
      - ZTA
  /access/relay/{uuid}:
    delete:
      description: 删除ZTA的relay
      parameters:
      - description: uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: DelRelay
      tags:
      - ZTA
...
  /sysytem/oauth2:
    get:
      description: 获取ZTA的Oauth2
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: ListOauth2
      tags:
      - ZTA Oauth2
    post:
      consumes:
      - application/json
      description: 新增ZTA的Oauth2
      parameters:
      - description: 新增ZTA的Oauth2
        in: body
        name: Oauth2
        required: true
        schema:
          $ref: '#/definitions/mparam.AddOauth2'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: AddOauth2
      tags:
      - ZTA Oauth2
    put:
      consumes:
      - application/json
      description: 修改ZTA的Oauth2
      parameters:
      - description: 修改ZTA的Oauth2
        in: body
        name: Oauth2
        required: true
        schema:
          $ref: '#/definitions/mparam.EditOauth2'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: EditOauth2
      tags:
      - ZTA Oauth2
  /sysytem/oauth2/{id}:
    delete:
      description: 删除ZTA的Oauth2
      parameters:
      - description: 主键ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Res'
      summary: DelOauth2
      tags:
      - ZTA Oauth2
swagger: "2.0"